+++ demo/demo.component.ts
@@ -4,7 +4,7 @@ import { Component, ViewEncapsulation, OnInit, OnDestroy } from '@angular/core';
   selector: 'mwl-demo-app',
   encapsulation: ViewEncapsulation.None,
   template: `
-    
+
     <div class="gauges-container">
 
       <mwl-gauge
@@ -60,23 +60,35 @@ import { Component, ViewEncapsulation, OnInit, OnDestroy } from '@angular/core';
         [max]="100"
         [dialStartAngle]="90.01"
         [dialEndAngle]="89.99"
-        [radius]="150"
+        [dialRadius]="10"
         [showValue]="false"
         [value]="gaugeValues[6]"
         [animated]="true"
         [animationDuration]="1">
       </mwl-gauge>
-      
+
+      <mwl-gauge
+        class="seven"
+        [max]="100"
+        [dialStartAngle]="-90"
+        [dialEndAngle]="-90.001"
+        [showValue]="false"
+        [value]="gaugeValues[6]"
+        [animated]="true"
+        [animationDuration]="1">
+      </mwl-gauge>
+
     </div>
 
   `,
-  styles: [`
+  styles: [
+    `
     body {
       background-color: rgba(0,0,0,0.8);
       color: #999;
       font-family: Hevletica, sans-serif;
     }
-    
+
     .gauges-container {
       display: flex;
     }
@@ -92,32 +104,32 @@ import { Component, ViewEncapsulation, OnInit, OnDestroy } from '@angular/core';
     }
     mwl-gauge > .gauge > .dial {
       stroke: #334455;
-      stroke-width: 20;
+      stroke-width: 2;
       fill: rgba(0,0,0,0);
     }
     mwl-gauge > .gauge > .value {
       stroke: rgb(47, 227, 255);
-      stroke-width: 20;
+      stroke-width: 2;
       fill: rgba(0,0,0,0);
     }
     mwl-gauge > .gauge > .value-text {
       fill: rgb(47, 227, 255);
       font-family: sans-serif;
       font-weight: bold;
-      font-size: 8em;
+      font-size: 0.8em;
     }
-    
+
     /* ------- Alternate Style ------- */
     mwl-gauge.two {
     }
     mwl-gauge.two > .gauge > .dial {
       stroke: #334455;
-      stroke-width: 100;
+      stroke-width: 10;
     }
     mwl-gauge.two > .gauge > .value {
       stroke: orange;
       stroke-dasharray: none;
-      stroke-width: 130;
+      stroke-width: 13;
     }
     mwl-gauge.two > .gauge > .value-text {
       fill: orange;
@@ -128,25 +140,25 @@ import { Component, ViewEncapsulation, OnInit, OnDestroy } from '@angular/core';
     }
     mwl-gauge.three > .gauge > .dial {
       stroke: #334455;
-      stroke-width: 20;
+      stroke-width: 2;
     }
     mwl-gauge.three > .gauge > .value {
       stroke: #C9DE3C;
-      stroke-width: 50;
+      stroke-width: 5;
     }
     mwl-gauge.three > .gauge > .value-text {
       fill: #C9DE3C;
     }
-    
+
     /* ----- Alternate Style ----- */
     mwl-gauge.four > .gauge > .dial {
       stroke: #334455;
-      stroke-width: 50;
+      stroke-width: 5;
     }
     mwl-gauge.four > .gauge > .value {
       stroke: #BE80FF;
       stroke-dasharray: none;
-      stroke-width: 50;
+      stroke-width: 5;
     }
     mwl-gauge.four > .gauge > .value-text {
       fill: #BE80FF;
@@ -155,37 +167,48 @@ import { Component, ViewEncapsulation, OnInit, OnDestroy } from '@angular/core';
     /* ----- Alternate Style ----- */
     mwl-gauge.five > .gauge > .dial {
       stroke: #334455;
-      stroke-width: 50;
+      stroke-width: 5;
     }
     mwl-gauge.five > .gauge > .value {
       stroke: #F8774B;
-      stroke-dasharray: none;
-      stroke-width: 50;
+      stroke-dasharray: 25 1;
+      stroke-width: 5;
     }
     mwl-gauge.five > .gauge > .value-text {
       fill: #F8774B;
-      font-size: 7em;
+      font-size: 0.7em;
     }
-    
+
     /* ----- Alternate Style ----- */
     mwl-gauge.six > .gauge > .dial {
       stroke: #334455;
-      stroke-dasharray: 104 14;
-      stroke-width: 70;
+      fill: #334455;
+      stroke-width: 20;
     }
     mwl-gauge.six > .gauge > .value {
       stroke: #FF6DAF;
-      stroke-dasharray: 104 14;
-      stroke-width: 70;
+      stroke-width: 20;
     }
     mwl-gauge.six > .gauge > .value-text {
       fill: #FF6DAF;
-      font-size: 7em;
+      font-size: 0.7em;
     }
-  `]
+
+    mwl-gauge.seven > .gauge > .dial {
+      stroke: transparent;
+      stroke-width: 5;
+      transform: scale(0.9,0.9) translate3d(5.5px, 5.5px, 0);
+      fill: rgba(148, 112, 57, 0.42);
+    }
+    mwl-gauge.seven > .gauge > .value {
+      stroke: #F8774B;
+      stroke-dasharray: none;
+      stroke-width: 5;
+    }
+  `
+  ]
 })
 export class DemoComponent implements OnInit, OnDestroy {
-
   percentageValue: (value: number) => string;
 
   gaugeValues: any = {
@@ -194,27 +217,28 @@ export class DemoComponent implements OnInit, OnDestroy {
     3: 50,
     4: 50,
     5: 50,
-    6: 50
+    6: 50,
+    7: 50
   };
 
-  interval: number;
+  interval: any;
 
   constructor() {
     this.percentageValue = function(value: number): string {
-      return `${Math.round(value)} / ${this.max}`;
+      return `${Math.round(value)} / ${this['max']}`;
     };
   }
 
   ngOnInit(): void {
-
-    const updateValues: Function = (): void => {
-      this.gaugeValues =  {
+    const updateValues = (): void => {
+      this.gaugeValues = {
         1: Math.round(Math.random() * 100),
         2: Math.round(Math.random() * 100),
         3: Math.round(Math.random() * 100),
         4: Math.round(Math.random() * 100),
         5: Math.round(Math.random() * 200),
-        6: Math.round(Math.random() * 100)
+        6: Math.round(Math.random() * 100),
+        7: Math.round(Math.random() * 100)
       };
     };
 
@@ -222,11 +246,9 @@ export class DemoComponent implements OnInit, OnDestroy {
 
     this.interval = setInterval(updateValues, INTERVAL);
     updateValues();
-
   }
 
   ngOnDestroy(): void {
     clearInterval(this.interval);
   }
-
 }