+++ karma.conf.ts
@@ -10,7 +10,7 @@ export default config => {
 
     // frameworks to use
     // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
-    frameworks: ['jasmine'],
+    frameworks: ['mocha'],
 
     // list of files / patterns to load in the browser
     files: [
@@ -32,16 +32,28 @@ export default config => {
           test: /\.ts$/,
           loader: 'tslint-loader',
           exclude: /node_modules/,
-          enforce: 'pre'
+          enforce: 'pre',
+          options: {
+            emitErrors: config.singleRun,
+            failOnHint: false
+          }
         }, {
           test: /\.ts$/,
-          loader: 'awesome-typescript-loader',
-          exclude: /node_modules/
+          loader: 'ts-loader',
+          exclude: /node_modules/,
+          options: {
+            compilerOptions: {
+              module: 'es2015'
+            }
+          }
         }, {
           test: /src\/.+\.ts$/,
           exclude: /(node_modules|\.spec\.ts$)/,
           loader: 'istanbul-instrumenter-loader',
-          enforce: 'post'
+          enforce: 'post',
+          options: {
+            esModules: true
+          }
         }]
       },
       plugins: [
@@ -49,19 +61,13 @@ export default config => {
           filename: null,
           test: /\.(ts|js)($|\?)/i
         }),
-        new webpack.LoaderOptionsPlugin({
-          options: {
-            tslint: {
-              emitErrors: config.singleRun,
-              failOnHint: false
-            }
-          }
-        }),
         new webpack.ContextReplacementPlugin(
-          /angular(\\|\/)core(\\|\/)@angular/,
+          /angular(\\|\/)core(\\|\/)esm5/,
           path.join(__dirname, 'src')
         ),
-        ...(config.singleRun ? [new webpack.NoEmitOnErrorsPlugin()] : [])
+        ...(config.singleRun ? [
+          new webpack.NoEmitOnErrorsPlugin()
+        ] : [])
       ]
     },
 
@@ -76,6 +82,10 @@ export default config => {
       }
     },
 
+    mime: {
+      'text/x-typescript': ['ts']
+    },
+
     // test results reporter to use
     // possible values: 'dots', 'progress'
     // available reporters: https://npmjs.org/browse/keyword/karma-reporter
@@ -87,17 +97,7 @@ export default config => {
 
     // start these browsers
     // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
-    browsers: ['PhantomJS'],
-
-    phantomjsLauncher: {
-      // Have phantomjs exit if a ResourceError is encountered (useful if karma exits without killing phantom)
-      exitOnResourceError: true
-    },
-
-    browserConsoleLogOptions: {
-      terminal: true,
-      level: 'log'
-    }
+    browsers: ['ChromeHeadless']
 
   });