+++ src/gauge.module.ts
@@ -2,7 +2,9 @@ import { NgModule, InjectionToken, ModuleWithProviders } from '@angular/core';
 import { GaugeComponent } from './gauge.component';
 import { GaugeDefaults, GaugeOptions } from './gauge-defaults.service';
 
-export const USER_DEFAULTS: InjectionToken<string> = new InjectionToken('gauge defaults');
+export const USER_DEFAULTS: InjectionToken<string> = new InjectionToken(
+  'gauge defaults'
+);
 
 export function defaultsFactory(userDefaults: GaugeOptions): GaugeDefaults {
   const defaults: GaugeDefaults = new GaugeDefaults();
@@ -11,27 +13,24 @@ export function defaultsFactory(userDefaults: GaugeOptions): GaugeDefaults {
 }
 
 @NgModule({
-  declarations: [
-    GaugeComponent
-  ],
-  exports: [
-    GaugeComponent
-  ]
+  declarations: [GaugeComponent],
+  exports: [GaugeComponent]
 })
 export class GaugeModule {
-
   static forRoot(userDefaults: GaugeOptions = {}): ModuleWithProviders {
     return {
       ngModule: GaugeModule,
-      providers: [{
-        provide: USER_DEFAULTS,
-        useValue: userDefaults
-      }, {
-        provide: GaugeDefaults,
-        useFactory: defaultsFactory,
-        deps: [USER_DEFAULTS]
-      }]
+      providers: [
+        {
+          provide: USER_DEFAULTS,
+          useValue: userDefaults
+        },
+        {
+          provide: GaugeDefaults,
+          useFactory: defaultsFactory,
+          deps: [USER_DEFAULTS]
+        }
+      ]
     };
   }
-
 }