+++ webpack.config.ts
@@ -1,6 +1,7 @@
 import * as webpack from 'webpack';
 import * as path from 'path';
 import * as HtmlWebpackPlugin from 'html-webpack-plugin';
+import * as ForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
 
 const IS_PROD: boolean = process.argv.indexOf('-p') > -1;
 
@@ -15,11 +16,21 @@ export default {
       test: /\.ts$/,
       loader: 'tslint-loader?emitErrors=false&failOnHint=false',
       exclude: /node_modules/,
-      enforce: 'pre'
+      enforce: 'pre',
+      options: {
+        emitErrors: false,
+        failOnHint: false
+      }
     }, {
       test: /\.ts$/,
-      loader: 'awesome-typescript-loader',
-      exclude: /node_modules/
+      loader: 'ts-loader',
+      exclude: /node_modules/,
+      options: {
+        transpileOnly: !IS_PROD,
+        compilerOptions: {
+          module: 'es2015'
+        }
+      }
     }]
   },
   resolve: {
@@ -32,12 +43,18 @@ export default {
     historyApiFallback: true
   },
   plugins: [
-    ...(IS_PROD ? [] : [new webpack.HotModuleReplacementPlugin()]),
+    ...(IS_PROD ? [] : [
+      new webpack.HotModuleReplacementPlugin(),
+      new ForkTsCheckerWebpackPlugin({
+        watch: ['./src', './demo'],
+        formatter: 'codeframe'
+      })
+    ]),
     new webpack.DefinePlugin({
       ENV: JSON.stringify(IS_PROD ? 'production' : 'development')
     }),
     new webpack.ContextReplacementPlugin(
-      /angular(\\|\/)core(\\|\/)@angular/,
+      /angular(\\|\/)core(\\|\/)esm5/,
       path.join(__dirname, 'src')
     ),
     new HtmlWebpackPlugin({